third commit

+++ flask_app/controllers/trees.py
@@ -1,5 +1,5 @@
 from flask_app import app
-from flask import render_template, redirect, request, session, Flask, flash, url_for
+from flask import render_template, redirect, request, session, Flask
 from flask_app.models import tree, picture
 from werkzeug.utils import secure_filename
 
@@ -9,6 +9,7 @@ ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg'}
 
 @app.post('/create/tree')
 def create_new_tree():
+    if 'user_id' not in session: return redirect('/')
     tree.Tree.create_new_tree(request.form)
     return redirect('/admin')
 
@@ -47,6 +48,7 @@ def update_one_tree():
 # Delete Tree Controller
 @app.post('/tree/delete')
 def delete_tree():
+    if 'user_id' not in session: return redirect('/')
     tree.Tree.delete_tree(request.form)
     return redirect('/admin')
 
@@ -55,34 +57,8 @@ def delete_tree():
 @app.post('/species/search')
 def search_by_common_name():
     if 'user_id' not in session: return redirect('/')
+    print(request.form)
     tree_name = request.form['name']
     tree_data = tree.Tree.get_tree_by_multiple_varieables(tree_name)
     print(tree_data, "Tree DATA")
     return redirect(f'/species/{tree_data}')
-# Notes:
-# 1 - Use meaningful names
-# 2 - Do not overwrite function names
-# 3 - No matchy, no worky
-# 4 - Use consistent naming conventions 
-# 5 - Keep it clean
-# 6 - Test every little line before progressing
-# 7 - READ ERROR MESSAGES!!!!!!
-# 8 - Error messages are found in the browser and terminal
-
-
-
-
-# How to use path variables:
-# @app.route('/<int:id>')                                   The variable must be in the path within angle brackets
-# def index(id):                                            It must also be passed into the function as an argument/parameter
-#     user_info = user.User.get_user_by_id(id)              The it will be able to be used within the function for that route
-#     return render_template('index.html', user_info)
-
-# Converter -	Description
-# string -	Accepts any text without a slash (the default).
-# int -	Accepts integers.
-# float -	Like int but for floating point values.
-# path 	-Like string but accepts slashes.
-
-# Render template is a function that takes in a template name in the form of a string, then any number of named arguments containing data to pass to that template where it will be integrated via the use of jinja
-# Redirect redirects from one route to another, this should always be done following a form submission. Don't render on a form submission.
